refactor: move summary renderer next to the setup command
The new `summary.ts` module is only consumed by `commands/setup/index.ts`
and its detection helpers (framework version, package manager) are
setup-specific. Living under `lib/` implied it was meant for reuse by
`apply`/`plan`/`doctor` — it isn't. Relocate to
`apps/cli/src/commands/setup/summary.ts` so the file is collocated with
its sole caller; if a second command ever wants a similar end-of-command
report, the generic `renderSummary`/`Row`/`Section` pieces can be
hoisted back to `lib/` at that point.
