+++ apps/cli/src/commands/setup/summary.ts
